PygcurseSurface.input() now passes its inputObj to its callbackFn().
My application needed access to the inputObj in use, and had to get at
it by use of hackish stack manipulation. PygcurseSurface.input() now
calls callbackFn() with inputObj as a parameter. This is designed to be
backwards-compatible with older Pygcurse users; if it turns out that
callbackFn() did not expect a parameter, we simply call it without the
parameter.

I have also moved the call to callbackFn() to before the event loop;
this way callbackFn() has access to pygame events before the event loop
flushes the event queue.